Update README.md
Added in with a successful photo to prove that after creating a  tag, tasking, and uploading a file successfully, then you can use the Secure Notes Feature.  simply put it needs to initialize random KEK and a DEK to begin randomly generating another one.  At first I wanted the user to select which keys in the database to encrypt with but then thought that would be bad practice so it always generates a random key no matter what, even if you select a different key.